pyproject.toml
[project]
name = "energy-consumption-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "xgboost>=3.2.0",
]
